requirements.txt
# Core data processing
pandas>=1.5.0
numpy>=1.21.0
pyarrow>=10.0.0
polars>=0.19.0

# AWS and cloud services
boto3>=1.26.0
botocore>=1.29.0
s3fs>=2023.1.0

# Database connectivity
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pymysql>=1.0.0
pyodbc>=4.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0
tqdm>=4.64.0

# Date and time handling
pytz>=2022.1
python-dateutil>=2.8.0

# Logging and monitoring
structlog>=22.1.0

# HTTP requests
requests>=2.28.0
urllib3>=1.26.0

# Parallel processing
joblib>=1.2.0
multiprocess>=0.70.0

# Scientific computing (optional, for advanced analytics)
scipy>=1.9.0
scikit-learn>=1.1.0

# Data validation
marshmallow>=3.17.0
cerberus>=1.3.0

# Progress bars and CLI
rich>=12.0.0
typer>=0.7.0

# File format support
openpyxl>=3.0.0
xlrd>=2.0.0
feather-format>=0.4.0

# Time series and statistics
statsmodels>=0.13.0

# Environment and system
psutil>=5.9.0

# Development and testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
awswrangler>=3.12.1
geopandas>=1.1.1
geopy>=2.4.1
pyathena>=3.15.0
rpy2>=3.6.1

# Install R (Ubuntu/Debian)
# sudo apt-get install r-base r-base-dev